+++ src/main/java/io/github/astrapi69/net/ip/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides classes for resolve ip addresses
+ */
+package io.github.astrapi69.net.ip;