+++ src/main/java/io/github/astrapi69/net/properties/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This package provides classes for the keys for properties of the network
+ */
+package io.github.astrapi69.net.properties;